Added tests and refactoring

+++ src/main/java/com/shop/demoshop/repositories/ProductRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-    List<Product> findBySubscribers_Id(long id);
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+    List<Product> findBySubscribers_Id(int id);
 }
